pyproject.toml
[project]
name = "personalised-career-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "gradio>=5.36.2",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "notebook>=7.4.5",
    "numpy>=2.3.2",
    "openai>=1.95.1",
    "openai-agents>=0.1.0",
    "pandas>=2.3.2",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
]
